uptime コマンドに 2 引数形式を追加し、GetTickCount64() へ移行
- uptime コマンドに 2 引数形式（第1引数: 秒、第2引数: ミリ秒）を追加
- GetTickCount()/DWORD を GetTickCount64()/ULONGLONG に置き換え（49 日オーバーフロー対策）
- 英日ドキュメントを更新
- テストスクリプト追加 (uptime, wait_timeout_connected)
- int 型への統一

+++ teraterm/common/tttypes.h
@@ -869,7 +869,7 @@ typedef struct {
 
 	void *NotifyIcon;
 
-	DWORD ConnectedTime;
+	ULONGLONG ConnectedTime;
 
 	void (*Log1Byte)(BYTE b);
 	void (*Log1Bin)(BYTE b);
